Remove old sample storage, replace with cpp
