+++ modyn/common/trigger_sample/include/trigger_sample_storage.hpp
@@ -3,20 +3,20 @@
 #include <fstream>
 #include <vector>
 
-namespace modyn::common::trigger_storage_cpp {
+namespace modyn::common::trigger_sample_storage {
 
 uint64_t get_num_samples_in_file_impl(const char* filename);
 
 void* get_worker_samples_impl(const char* folder, uint64_t* size, const char* pattern, uint64_t start_index,
-                              size_t worker_subset_size);
+                              uint64_t worker_subset_size);
 void* get_all_samples_impl(const char* folder, uint64_t* size, const char* pattern);
 
 void* parse_file_impl(const char* filename, uint64_t* size);
 
-void write_file_impl(const char* filename, const void* data, size_t data_offset, size_t data_length, const char* header,
-                     size_t header_length);
-void write_files_impl(const char* filenames[], const void* data, std::size_t data_lengths[], const char* headers[],
-                      std::size_t header_length, std::size_t num_files);
+void write_file_impl(const char* filename, const void* data, uint64_t data_offset, uint64_t data_length,
+                     const char* header, uint64_t header_length);
+void write_files_impl(const char* filenames[], const void* data, std::uint64_t data_lengths[], const char* headers[],
+                      std::uint64_t header_length, std::uint64_t num_files);
 
 void release_data_impl(void* data);
 
@@ -24,9 +24,9 @@ bool parse_file_subset(const char* filename, std ::vector<char>& char_vector, ui
                        uint64_t end_index);
 std::vector<std::string> get_matching_files(const char* folder, const char* pattern);
 
-std::size_t read_data_size_from_header(std::ifstream& file);
+std::uint64_t read_data_size_from_header(std::ifstream& file);
 int read_magic(std::ifstream& file);
 
 std::ifstream open_file(const char* filename);
 std::ofstream open_file_write(const char* filename);
-}  // namespace modyn::common::trigger_storage_cpp
+}  // namespace modyn::common::trigger_sample_storage
