+++ modyn/common/trigger_sample/trigger_sample_storage.py
@@ -1,12 +1,49 @@
+import ctypes
 import logging
 import os
 import sys
+import typing
 from pathlib import Path
+from sys import platform
 
 import numpy as np
-from modyn.utils import flatten, get_partition_for_worker
+from numpy.ctypeslib import ndpointer
+
+from .utils import get_partition_for_worker
 
 logger = logging.getLogger(__name__)
+
+
+class ArrayWrapper:
+    def __init__(self, array: np.ndarray, f_release: typing.Callable) -> None:
+        self.array = array
+        self.f_release = f_release
+
+    def __len__(self) -> int:
+        return self.array.size
+
+    def __getitem__(self, key: typing.Any) -> typing.Any:
+        return self.array.__getitem__(key)
+
+    def __str__(self) -> str:
+        return self.array.__str__()
+
+    def __del__(self) -> None:
+        self.f_release(self.array)
+
+    def __eq__(self, other: typing.Any) -> typing.Any:
+        return self.array == other
+
+    def __ne__(self, other: typing.Any) -> typing.Any:
+        return self.array == other
+
+    @property
+    def dtype(self) -> np.dtype:
+        return self.array.dtype
+
+    @property
+    def shape(self) -> tuple:
+        return self.array.shape
 
 
 class TriggerSampleStorage:
@@ -21,16 +58,91 @@ class TriggerSampleStorage:
     sample weight separated by a comma.
     """
 
-    def __init__(
-        self,
-        trigger_sample_directory: str,
-    ):
+    @staticmethod
+    def _get_library_path() -> Path:
+        if platform == "darwin":
+            library_filename = "libtrigger_sample_storage.dylib"
+        else:
+            library_filename = "libtrigger_sample_storage.so"
+
+        return TriggerSampleStorage._get_build_path() / library_filename
+
+    @staticmethod
+    def _get_build_path() -> Path:
+        return Path(__file__).parent.parent.parent.parent / "libbuild"
+
+    @staticmethod
+    def _ensure_library_present() -> None:
+        path = TriggerSampleStorage._get_library_path()
+        if not path.exists():
+            raise RuntimeError(f"Cannot find {TriggerSampleStorage.__name__} library at {path}")
+
+    def __init__(self, trigger_sample_directory: str = "sample_dir") -> None:
+        TriggerSampleStorage._ensure_library_present()
+        self.extension = ctypes.CDLL(str(TriggerSampleStorage._get_library_path()))
+
         self.trigger_sample_directory = trigger_sample_directory
         if not Path(self.trigger_sample_directory).exists():
             Path(self.trigger_sample_directory).mkdir(parents=True, exist_ok=True)
             logger.info(f"Created the trigger sample directory {self.trigger_sample_directory}.")
         if sys.maxsize < 2**63 - 1:
             raise RuntimeError("Modyn Selector Implementation requires a 64-bit system.")
+
+        # We define a return object here that can infer its size when being parsed by ctypes
+        self.data_pointer = ndpointer(dtype=[("f0", "<i8"), ("f1", "<f8")], ndim=1, shape=(1,), flags="C_CONTIGUOUS")
+        self.data_pointer._shape_ = property(lambda self: self.shape_val[0])
+
+        self._get_num_samples_in_file_impl = self.extension.get_num_samples_in_file
+        self._get_num_samples_in_file_impl.argtypes = [ctypes.POINTER(ctypes.c_char)]
+        self._get_num_samples_in_file_impl.restype = ctypes.c_uint64
+
+        self._write_file_impl = self.extension.write_file
+        self._write_file_impl.argtypes = [
+            ctypes.POINTER(ctypes.c_char),
+            ndpointer(flags="C_CONTIGUOUS"),
+            ctypes.c_uint64,
+            ctypes.c_uint64,
+            ctypes.POINTER(ctypes.c_char),
+            ctypes.c_uint64,
+        ]
+        self._write_file_impl.restype = None
+
+        self._write_files_impl = self.extension.write_files
+        self._write_files_impl.argtypes = [
+            ctypes.POINTER(ctypes.c_char_p),
+            ndpointer(flags="C_CONTIGUOUS"),
+            ctypes.POINTER(ctypes.c_uint64),
+            ctypes.POINTER(ctypes.c_char_p),
+            ctypes.c_uint64,
+            ctypes.c_uint64,
+        ]
+        self._write_files_impl.restype = None
+
+        self._get_all_samples_impl = self.extension.get_all_samples
+        self._get_all_samples_impl.argtypes = [
+            ctypes.POINTER(ctypes.c_char),
+            ctypes.POINTER(ctypes.c_uint64),
+            ctypes.POINTER(ctypes.c_char),
+        ]
+        self._get_all_samples_impl.restype = self.data_pointer
+
+        self._get_worker_samples_impl = self.extension.get_worker_samples
+        self._get_worker_samples_impl.argtypes = [
+            ctypes.POINTER(ctypes.c_char),
+            ctypes.POINTER(ctypes.c_uint64),
+            ctypes.POINTER(ctypes.c_char),
+            ctypes.c_uint64,
+            ctypes.c_uint64,
+        ]
+        self._get_worker_samples_impl.restype = self.data_pointer
+
+        self._parse_file_impl = self.extension.parse_file
+        self._parse_file_impl.argtypes = [ctypes.POINTER(ctypes.c_char), ctypes.POINTER(ctypes.c_uint64)]
+        self._parse_file_impl.restype = self.data_pointer
+
+        self._release_data_impl = self.extension.release_data
+        self._release_data_impl.argtypes = [ndpointer(dtype=[("f0", "<i8"), ("f1", "<f8")], flags="C_CONTIGUOUS")]
+        self._release_data_impl.restype = None
 
     def get_trigger_samples(
         self,
@@ -40,7 +152,7 @@ class TriggerSampleStorage:
         retrieval_worker_id: int = -1,
         total_retrieval_workers: int = -1,
         num_samples_trigger_partition: int = -1,
-    ) -> list[tuple[int, float]]:
+    ) -> ArrayWrapper:
         """
         Return the trigger samples for the given pipeline id, trigger id and partition id.
 
@@ -84,7 +196,7 @@ class TriggerSampleStorage:
         retrieval_worker_id: int,
         total_retrieval_workers: int,
         num_samples_trigger_partition: int,
-    ) -> list[tuple[int, float]]:
+    ) -> ArrayWrapper:
         """
         Return the trigger samples for the given pipeline id, trigger id and partition id that are assigned to the
         retrieval worker.
@@ -101,64 +213,16 @@ class TriggerSampleStorage:
             retrieval_worker_id, total_retrieval_workers, num_samples_trigger_partition
         )
 
-        current_index = 0
+        folder = ctypes.c_char_p(str(self.trigger_sample_directory).encode("utf-8"))
+        size = (ctypes.c_uint64 * 1)()
+        self.data_pointer.shape_val = size
+        pattern = ctypes.c_char_p(f"{pipeline_id}_{trigger_id}_{partition_id}_".encode("utf-8"))
 
-        triple_list: list[tuple[Path, int, int]] = []
-        for file in sorted(os.listdir(self.trigger_sample_directory)):
-            if file.startswith(f"{pipeline_id}_{trigger_id}_{partition_id}_"):
-                file_path = Path(self.trigger_sample_directory) / file
-                if current_index >= start_index + worker_subset_size:
-                    #  We have already retrieved all the samples for the worker
-                    break
-                num_samples_in_file = self._get_num_samples_in_file(file_path)
-                if current_index + num_samples_in_file <= start_index:
-                    # The samples in the file are before the samples for the worker
-                    current_index += num_samples_in_file
-                    continue
-                if current_index + num_samples_in_file < start_index + worker_subset_size:
-                    # The head of samples for the worker are in the file, either partially from
-                    # start_index - current_index to the end of the file if start_index > current_index
-                    # or completely from 0 to the end of the file.
-                    # Because the end index is exclusive, we compare < instead of <= otherwise we would retrieve
-                    # one more sample than we should
-                    triple_list.append(
-                        (
-                            file_path,
-                            start_index - current_index if start_index - current_index >= 0 else 0,
-                            num_samples_in_file,
-                        )
-                    )
-                    current_index += num_samples_in_file
-                    continue
-                # We are at the tail of the file and the samples for the worker are in the file, either from
-                #  the beginning if start_index - current_index < 0 or from start_index - current_index if the
-                #  tail is in the same file as the head
-                triple_list.append(
-                    (
-                        file_path,
-                        start_index - current_index if start_index - current_index >= 0 else 0,
-                        start_index + worker_subset_size - current_index,
-                    )
-                )
-                break
+        data = self._get_worker_samples_impl(folder, size, pattern, start_index, worker_subset_size).reshape(-1)
+        result = ArrayWrapper(data, self._release_data_impl)
+        return result
 
-        # We need to flatten the list of lists of np arrays and then reshape it to get the list of tuples
-        return [
-            (int(key), float(weight))  # type: ignore
-            for (key, weight) in map(
-                tuple,  # type: ignore
-                np.array(
-                    flatten(
-                        [
-                            self._parse_file_subset(file_path, start_index, end_index)
-                            for file_path, start_index, end_index in triple_list
-                        ]
-                    )
-                ),
-            )
-        ]
-
-    def _get_all_samples(self, pipeline_id: int, trigger_id: int, partition_id: int) -> list[tuple[int, float]]:
+    def _get_all_samples(self, pipeline_id: int, trigger_id: int, partition_id: int) -> ArrayWrapper:
         """
         Return all the samples for the given pipeline id, trigger id and partition id.
 
@@ -168,29 +232,17 @@ class TriggerSampleStorage:
         :return: the trigger samples
         """
 
-        return [
-            (int(key), float(weight))  # type: ignore
-            for (key, weight) in map(
-                tuple,  # type: ignore
-                np.array(
-                    flatten(
-                        [
-                            self._parse_file(Path(self.trigger_sample_directory) / file)
-                            for file in sorted(os.listdir(self.trigger_sample_directory))
-                            if file.startswith(f"{pipeline_id}_{trigger_id}_{partition_id}_")
-                        ]
-                    )
-                ),
-            )
-        ]
+        folder = ctypes.c_char_p(str(self.trigger_sample_directory).encode("utf-8"))
+        size = (ctypes.c_uint64 * 1)()
+        self.data_pointer.shape_val = size
+        pattern = ctypes.c_char_p(f"{pipeline_id}_{trigger_id}_{partition_id}_".encode("utf-8"))
+
+        data = self._get_all_samples_impl(folder, size, pattern).reshape(-1)
+        result = ArrayWrapper(data, self._release_data_impl)
+        return result
 
     def save_trigger_sample(
-        self,
-        pipeline_id: int,
-        trigger_id: int,
-        partition_id: int,
-        trigger_samples: np.ndarray,
-        insertion_id: int,
+        self, pipeline_id: int, trigger_id: int, partition_id: int, trigger_samples: np.ndarray, insertion_id: int
     ) -> None:
         """
         Save the trigger samples for the given pipeline id, trigger id and partition id.
@@ -214,55 +266,33 @@ class TriggerSampleStorage:
 
         self._write_file(samples_file, trigger_samples)
 
-    def _write_file(self, file_path: Path, trigger_samples: np.ndarray) -> None:
-        """Write the trigger samples to the given file.
-
-        Args:
-            file_path (str): File path to write to.
-            trigger_samples (list[tuple[int, float]]): List of trigger samples.
+    def save_trigger_samples(
+        self, pipeline_id: int, trigger_id: int, partition_id: int, trigger_samples: np.ndarray, data_lengths: list
+    ) -> None:
         """
-        np.save(file_path, trigger_samples, allow_pickle=False, fix_imports=False)
+        Save the trigger samples for the given pipeline id, trigger id and partition id
+        to multiple files.
 
-    def _parse_file(self, file_path: Path) -> np.ndarray:
-        """Parse the given file and return the samples.
-
-        Args:
-            file_path (str): File path to parse.
-
-        Returns:
-            list[tuple[int, float]]: List of trigger samples.
+        :param pipeline_id: the id of the pipeline
+        :param trigger_id: the id of the trigger
+        :param partition_id: the id of the partition
+        :param trigger_samples: the trigger samples
+        :param data_lengths: the lengths of subarrays to write to files
         """
+        if trigger_samples.dtype != np.dtype("i8,f8"):
+            raise ValueError(f"Unexpected dtype: {trigger_samples.dtype}\nExpected: {np.dtype('i8,f8')}")
 
-        # When there are few samples, it may happen that some workers don't have any samples to store.
-        # Therefore, they do not write anything; you get an error if you try to read their file.
-        # This way, it returns an empty array if the worker has not written anything.
-        if not os.path.isfile(file_path):
-            return np.ndarray(0, dtype="i8,f8")
+        Path(self.trigger_sample_directory).mkdir(parents=True, exist_ok=True)
 
-        return np.load(file_path, allow_pickle=False, fix_imports=False)
+        samples_files = []
+        for i in range(len(data_lengths)):
+            samples_files.append(Path(self.trigger_sample_directory) / f"{pipeline_id}_{trigger_id}_{partition_id}_{i}")
+            assert not Path(samples_files[i]).exists(), (
+                f"Trigger samples file {samples_files[i]} already exists. "
+                f"Please delete it if you want to overwrite it."
+            )
 
-    def _parse_file_subset(self, file_path: Path, start_index: int, end_index: int) -> np.memmap:
-        """Parse the given file and return the samples. Only return samples between start_index
-           inclusive and end_index exclusive.
-
-        Args:
-            file_path (str): File path to parse.
-            end_index (int): The index of the last sample to return.
-
-        Returns:
-            list[tuple[int, float]]: List of trigger samples.
-        """
-        return np.load(file_path, allow_pickle=False, fix_imports=False, mmap_mode="r").take(
-            range(start_index, end_index), axis=0
-        )
-
-    def _get_num_samples_in_file(self, file_path: Path) -> int:
-        """Get the number of samples in the given file.
-
-        Args:
-            file_path (str): File path to parse.
-        """
-        return np.load(file_path, allow_pickle=False, fix_imports=False, mmap_mode="r").shape[0]
+        self._write_files(samples_files, trigger_samples, data_lengths)
 
     def get_file_path(self, pipeline_id: int, trigger_id: int, partition_id: int, worker_id: int) -> Path:
         return Path(self.trigger_sample_directory) / f"{pipeline_id}_{trigger_id}_{partition_id}_{worker_id}.npy"
@@ -272,8 +302,7 @@ class TriggerSampleStorage:
 
         return list(
             filter(
-                lambda file: file.startswith(f"{pipeline_id}_{trigger_id}_"),
-                os.listdir(self.trigger_sample_directory),
+                lambda file: file.startswith(f"{pipeline_id}_{trigger_id}_"), os.listdir(self.trigger_sample_directory)
             )
         )
 
@@ -294,3 +323,116 @@ class TriggerSampleStorage:
 
             for file in this_trigger_files:
                 os.remove(os.path.join(self.trigger_sample_directory, file))
+
+    def _parse_file(self, file_path: Path) -> np.ndarray:
+        """Parse the given file and return the samples.
+
+        Args:
+            file_path (str): File path to parse.
+
+        Returns:
+            np.ndarray: List of trigger samples.
+        """
+
+        file = ctypes.c_char_p(str(file_path).encode("utf-8"))
+        size = (ctypes.c_uint64 * 1)()
+        self.data_pointer.shape_val = size
+
+        data = self._parse_file_impl(file, size).reshape(-1)
+        result = ArrayWrapper(data, self._release_data_impl)
+
+        return result
+
+    def _get_num_samples_in_file(self, file_path: Path) -> int:
+        """Get the number of samples in the given file.
+
+        Args:
+            file_path (str): File path to parse.
+        """
+
+        file = ctypes.c_char_p(str(file_path).encode("utf-8"))
+        return self._get_num_samples_in_file_impl(file)
+
+    def _write_file(self, file_path: Path, trigger_samples: list) -> None:
+        """Write the trigger samples to the given file.
+
+        Args:
+            file_path (str): File path to write to.
+            trigger_samples (np.ndarray): List of trigger samples.
+        """
+
+        data = np.asanyarray(trigger_samples)
+
+        header = self._build_array_header(np.lib.format.header_data_from_array_1_0(data))
+
+        file = ctypes.c_char_p(str(file_path.with_suffix(".npy")).encode("utf-8"))
+        header = ctypes.c_char_p(header)
+
+        self._write_file_impl(file, data, 0, len(data), header, 128)
+
+    def _write_files(self, file_paths: list, trigger_samples: np.ndarray, data_lengths: list) -> None:
+        """Write the trigger samples to multiple files.
+
+        Args:
+            file_path (str): File path to write to.
+            trigger_samples (np.ndarray): List of trigger samples.
+            data_lengths (list): List of
+        """
+
+        data = np.asanyarray(trigger_samples)
+
+        raw_headers = []
+        length_sum = 0
+
+        # To comply with the numpy binary format, we need to start the file with a numpy array header.
+        # We call a modified version of the numpy header generating function and pass the resulting
+        # header strings on to C.
+        for data_length in data_lengths:
+            raw_headers.append(
+                self._build_array_header(
+                    np.lib.format.header_data_from_array_1_0(data[length_sum : length_sum + data_length])
+                )
+            )
+            length_sum += data_length
+
+        files = [ctypes.c_char_p(str(file_path.with_suffix(".npy")).encode("utf-8")) for file_path in file_paths]
+        headers = [ctypes.c_char_p(header) for header in raw_headers]
+
+        files_p = (ctypes.c_char_p * len(files))()
+        headers_p = (ctypes.c_char_p * len(files))()
+        data_lengths_p = (ctypes.c_uint64 * len(data_lengths))()
+
+        for i, _ in enumerate(files):
+            files_p[i] = files[i]
+            headers_p[i] = headers[i]
+            data_lengths_p[i] = data_lengths[i]
+
+        self._write_files_impl(files_p, data, data_lengths_p, headers_p, 128, len(data_lengths))
+
+    def _build_array_header(self, d: dict) -> str:
+        """Build the header for the array
+        Sourced from NumPy, modified version of _write_array_header:
+        https://github.com/numpy/numpy/blob/main/numpy/lib/format.py
+
+        Args:
+            d (dict): Dictionary of header items
+
+        Returns:
+            str: Header string
+        """
+        header = ["{"]
+        for key, value in sorted(d.items()):
+            # Need to use repr here, since we eval these when reading
+            header.append(f"'{key}': {repr(value)}, ")
+        header.append("}")
+        header = "".join(header)
+
+        # Add some spare space so that the array header can be modified in-place
+        # when changing the array size, e.g. when growing it by appending data at
+        # the end.
+        shape = d["shape"]
+        growth_axis_max_digits = 21
+        header += " " * ((growth_axis_max_digits - len(repr(shape[0]))) if len(shape) > 0 else 0)
+
+        header = np.lib.format._wrap_header_guess_version(header)
+        return header
