--- modyn/common/trigger_storage_cpp/__init__.py
@@ -1,10 +1,0 @@
-"""
-This module contains classes shared among several components.
-"""
-import os
-
-from .trigger_storage_cpp import TriggerStorageCPP as TriggerSampleStorage
-
-files = os.listdir(os.path.dirname(__file__))
-files.remove("__init__.py")
-__all__ = [f[:-3] for f in files if f.endswith(".py")]
