+++ modyn/tests/common/trigger_storage_cpp/test_trigger_storage_cpp.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import numpy as np
 import pytest
-from modyn.common.trigger_storage_cpp.trigger_storage_cpp import TriggerStorageCPP
+from modyn.common.trigger_sample.trigger_sample_storage import TriggerSampleStorage
 
 TMP_DIR = tempfile.mkdtemp()
 
@@ -20,9 +20,11 @@ def setup_and_teardown():
 
 
 def test_save_trigger_sample():
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(1, 2, 3, np.array([(1, 1.0), (2, 2.0)], dtype=np.dtype("i8,f8")), 4)
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
+        1, 2, 3, np.array([(1, 1.0), (2, 2.0)], dtype=np.dtype("i8,f8")), 4
+    )
 
-    samples = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3)
+    samples = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3)
 
     assert len(samples) == 2
     assert samples[0][0] == 1
@@ -32,11 +34,11 @@ def test_save_trigger_sample():
 
 
 def test_save_trigger_samples():
-    TriggerStorageCPP(TMP_DIR).save_trigger_samples(
+    TriggerSampleStorage(TMP_DIR).save_trigger_samples(
         1, 2, 3, np.array([(1, 1.0), (2, 2.0), (3, 3.0), (4, 4.0), (5, 5.0)], dtype=np.dtype("i8,f8")), [3, 2]
     )
 
-    samples = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3)
+    samples = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3)
 
     assert len(samples) == 5
     assert samples[0][0] == 1
@@ -52,29 +54,33 @@ def test_save_trigger_samples():
 
 
 def test_get_file_size():
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(1, 2, 3, np.array([(1, 1.0), (2, 2.0)], dtype=np.dtype("i8,f8")), 4)
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
+        1, 2, 3, np.array([(1, 1.0), (2, 2.0)], dtype=np.dtype("i8,f8")), 4
+    )
 
     file_path = Path(TMP_DIR) / "1_2_3_4.npy"
-    TriggerStorageCPP(TMP_DIR)._get_num_samples_in_file(file_path)
+    TriggerSampleStorage(TMP_DIR)._get_num_samples_in_file(file_path)
 
 
 def test_parse_file():
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(1, 2, 3, np.array([(1, 1.0), (2, 2.0)], dtype=np.dtype("i8,f8")), 4)
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
+        1, 2, 3, np.array([(1, 1.0), (2, 2.0)], dtype=np.dtype("i8,f8")), 4
+    )
     file_path = Path(TMP_DIR) / "1_2_3_4.npy"
 
-    samples = TriggerStorageCPP(TMP_DIR)._parse_file(file_path)
+    samples = TriggerSampleStorage(TMP_DIR)._parse_file(file_path)
     assert len(samples) == 2
     assert samples[0][0] == 1
     assert samples[0][1] == 1.0
     assert samples[1][0] == 2
     assert samples[1][1] == 2.0
 
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 3, np.array([(1, 1.0), (2, 2.0), (3, 3.0), (4, 4.0)], dtype=np.dtype("i8,f8")), 5
     )
     file_path = Path(TMP_DIR) / "1_2_3_5.npy"
 
-    samples = TriggerStorageCPP(TMP_DIR)._parse_file(file_path)
+    samples = TriggerSampleStorage(TMP_DIR)._parse_file(file_path)
     assert len(samples) == 4
     assert samples[0][0] == 1
     assert samples[0][1] == 1.0
@@ -87,16 +93,16 @@ def test_parse_file():
 
 
 def test_get_trigger_samples():
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 3, np.array([(1, 1.0), (2, 2.0), (3, 3.0), (4, 4.0)], dtype=np.dtype("i8,f8")), 4
     )
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 3, np.array([(3, 3.0), (4, 4.0), (5, 5.0), (6, 6.0)], dtype=np.dtype("i8,f8")), 5
     )
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 3, np.array([(7, 7.0), (8, 8.0), (8, 8.0), (8, 8.0)], dtype=np.dtype("i8,f8")), 6
     )
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 3, np.array([(9, 9.0), (10, 10.0), (11, 11.0), (12, 12.0)], dtype=np.dtype("i8,f8")), 7
     )
 
@@ -122,46 +128,46 @@ def test_get_trigger_samples():
         dtype=np.dtype("i8,f8"),
     )
 
-    assert (TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3) == expected_order).all()
+    assert (TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3) == expected_order).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 4, 16)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 4, 16)
     assert len(result) == 4
     assert (result == expected_order[0:4]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 1, 4, 16)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 1, 4, 16)
     assert len(result) == 4
     assert (result == expected_order[4:8]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 2, 4, 16)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 2, 4, 16)
     assert len(result) == 4
     assert (result == expected_order[8:12]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 3, 4, 16)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 3, 4, 16)
     assert len(result) == 4
     assert (result == expected_order[12:16]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 3, 16)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 3, 16)
     assert len(result) == 6
     assert (result == expected_order[0:6]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 2, 3, 16)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 2, 3, 16)
     assert len(result) == 5
     assert (result == expected_order[11:16]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 1, 16)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 1, 16)
     assert len(result) == 16
     assert (result == expected_order).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 9, 10, 16)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 9, 10, 16)
     assert len(result) == 1
     assert result == expected_order[15]
 
 
 def test_extended_get_trigger_samples():
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 3, np.array([(1, 1.0), (2, 2.0), (3, 3.0), (4, 4.0)], dtype=np.dtype("i8,f8")), 4
     )
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 3, np.array([(3, 3.0), (4, 4.0), (5, 5.0), (6, 6.0)], dtype=np.dtype("i8,f8")), 5
     )
 
@@ -169,7 +175,7 @@ def test_extended_get_trigger_samples():
         [(1, 1.0), (2, 2.0), (3, 3.0), (4, 4.0), (3, 3.0), (4, 4.0), (5, 5.0), (6, 6.0)], dtype=np.dtype("i8,f8")
     )
 
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 3, np.array([(7, 7.0), (8, 8.0), (9, 9.0), (10, 10.0), (11, 11.0)], dtype=np.dtype("i8,f8")), 6
     )
     expected_order = np.array(
@@ -191,67 +197,67 @@ def test_extended_get_trigger_samples():
         dtype=np.dtype("i8,f8"),
     )
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 2, 12)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 2, 12)
     assert len(result) == 6
     assert (result == expected_order[0:6]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 1, 3, 12)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 1, 3, 12)
     assert len(result) == 4
     assert (result == expected_order[4:8]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 2, 3, 12)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 2, 3, 12)
     assert len(result) == 4
     assert (result == expected_order[8:12]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 12, 12)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 12, 12)
     assert len(result) == 1
     assert result == expected_order[0]
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 1, 5, 12)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 1, 5, 12)
     assert len(result) == 3
     assert (result == expected_order[3:6]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 2, 5, 12)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 2, 5, 12)
     assert len(result) == 2
     assert (result == expected_order[6:8]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 3, 5, 12)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 3, 5, 12)
     assert len(result) == 2
     assert (result == expected_order[8:10]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 4, 5, 12)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 4, 5, 12)
     assert len(result) == 2
     assert (result == expected_order[10:12]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, -1, -1, 13)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, -1, -1, 13)
     assert len(result) == 13
     assert (result == expected_order).all()
 
-    TriggerStorageCPP(TMP_DIR).save_trigger_sample(
+    TriggerSampleStorage(TMP_DIR).save_trigger_sample(
         1, 2, 30, np.array([(1, 1.0), (2, 2.0), (3, 3.0), (4, 4.0)], dtype=np.dtype("i8,f8")), 6
     )
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 2, 13)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 0, 2, 13)
     assert len(result) == 7
     assert (result == expected_order[0:7]).all()
 
-    result = TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 30, -1, -1, 4)
+    result = TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 30, -1, -1, 4)
     assert len(result) == 4
     assert (result == expected_order[0:4]).all()
 
 
 def test_get_trigger_samples_no_file():
     with pytest.raises(ValueError):
-        TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3)
+        TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3)
 
 
 def test_get_trigger_samples_illegal_workers():
     with pytest.raises(AssertionError):
-        TriggerStorageCPP(TMP_DIR).get_trigger_samples(1, 2, 3, 0, -1, 2)
+        TriggerSampleStorage(TMP_DIR).get_trigger_samples(1, 2, 3, 0, -1, 2)
 
 
 def test_init_directory():
     os.rmdir(TMP_DIR)
-    _ = TriggerStorageCPP(TMP_DIR)
+    _ = TriggerSampleStorage(TMP_DIR)
 
     assert os.path.exists(TMP_DIR)
